Issue #10: Completed, buscador ainda non comprobado
