+++ project/frontend/app/src/main/AndroidManifest.xml
@@ -15,16 +15,16 @@
         android:theme="@style/Theme.Licopedia"
         tools:targetApi="31">
         <activity
+            android:name=".cocktails_fragment"
+            android:exported="false" />
+        <activity
             android:name=".MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
+
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
-        </activity>
-        <activity
-            android:name=".cocktails_activity"
-            android:exported="false">
         </activity>
     </application>
