+++ project/frontend/app/src/main/res/layout/activity_cocktails.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".cocktails_activity">
+    tools:context=".cocktails_fragment">
 
     <TextView
         android:id="@+id/text_view"
@@ -17,6 +17,25 @@
         app:layout_constraintLeft_toLeftOf="parent"
         app:layout_constraintRight_toRightOf="parent"/>
 
+    <EditText
+        android:id="@+id/searchCocktailByName"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:hint="Escribe el nombre del cocktail"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/cocktail_img" />
+
+    <ImageButton
+        android:background="@color/white"
+        android:id="@+id/searchButton"
+        android:layout_width="40dp"
+        android:layout_height="40dp"
+        android:scaleType="centerInside"
+        android:src="@drawable/search_icon"
+        app:layout_constraintLeft_toRightOf="@id/searchCocktailByName"
+        app:layout_constraintTop_toBottomOf="@+id/cocktail_img"
+        />
 
     <ImageButton
         android:id="@+id/profile_button"
@@ -54,7 +73,7 @@
         android:id="@+id/recyclerView"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        app:layout_constraintTop_toBottomOf="@id/cocktail_img"
+        app:layout_constraintTop_toBottomOf="@id/searchCocktailByName"
         app:layout_constraintLeft_toLeftOf="parent"
         app:layout_constraintRight_toRightOf="parent"
         />
